backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pymupdf>=1.25.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
python-multipart>=0.0.20
httpx>=0.28.0
pyyaml>=6.0.0
supabase>=2.0.0
python-jose[cryptography]>=3.5.0
gunicorn>=23.0.0
